update and add docker files
fix docker/base/ubuntu_1604_cuda90_cudnn7/Dockerfile

fix more on cuda9x docker files

fix docker/base/ubuntu_1604/Dockerfile

fix /docker/base/ubuntu_1804_cuda92_cudnn7/Dockerfile

fix gadgets/CMakeLists.txt for python

fix boost python check

fix for python3

update cuda versions

update ubuntu_1604_cuda80_cudnn6/7 for python3

fix for docker/base/ubuntu_1604_cuda80_cudnn6/Dockerfile docker/base/ubuntu_1604_cuda80_cudnn7/Dockerfile

fix boost python finding in docker

remove FindPython3 outputs

fix docker/base/ubuntu_1604/Dockerfile

make python_domo works in python2

update docker/push_base_images.sh

fix /FindPython3.cmake

+++ gadgets/python/gadgets/accumulate_and_recon.py
@@ -28,10 +28,10 @@ class AccumulateAndRecon(Gadget):
             eNy = self.enc.encodedSpace.matrixSize.y
             eNx = self.enc.encodedSpace.matrixSize.x
         
-            self.myBuffer = np.zeros((eNx/2,eNy,eNz,nslices,channels),dtype=np.complex64)
+            self.myBuffer = np.zeros(( int(eNx/2),eNy,eNz,nslices,channels),dtype=np.complex64)
 
         line_offset = self.enc.encodedSpace.matrixSize.y/2 - self.enc.encodingLimits.kspace_encoding_step_1.center             
-        self.myBuffer[:,acq.idx.kspace_encode_step_1+line_offset,acq.idx.kspace_encode_step_2,acq.idx.slice,:] = data
+        self.myBuffer[:,int(acq.idx.kspace_encode_step_1+line_offset), int(acq.idx.kspace_encode_step_2), int(acq.idx.slice),:] = data
 
         if (acq.flags & (1<<7)): #Is this the last scan in slice
             image = transform.transform_kspace_to_image(self.myBuffer,dim=(0,1,2))
